code/requirements.txt
aniso8601==7.0.0
annotated-types==0.7.0
anyio==4.4.0
async-exit-stack==1.0.1
asyncpg==0.29.0
bcrypt==4.1.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
graphene==2.1.9
graphql-core==3.2.3
graphql-relay==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
packaging==24.1
passlib==1.7.4
pipdeptree==2.23.1
promise==2.3
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic-sqlalchemy==0.0.9
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.7
rich==13.7.1
Rx==1.6.3
shellingham==1.5.4
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.37.2
strawberry-graphql==0.236.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
